requirements-dev.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
apify==3.0.3
apify_client==2.2.1
apify_shared==2.1.0
attrs==25.4.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
crawlee==1.0.4
cryptography==46.0.3
deprecation==2.1.0
fastapi==0.120.1
filelock==3.20.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
hyperframe==6.1.0
idna==3.11
impit==0.8.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lazy-object-proxy==1.12.0
mcp==1.19.0
more-itertools==10.8.0
multidict==6.7.0
packaging==25.0
postgrest==2.22.2
propcache==0.4.1
Protego==0.5.0
psutil==7.1.2
pycparser==2.23
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
pyee==13.0.0
PyJWT==2.10.1
python-dotenv==1.2.1
python-multipart==0.0.20
realtime==2.22.2
referencing==0.37.0
requests==2.32.5
requests-file==3.0.1
rpds-py==0.28.0
sniffio==1.3.1
sse-starlette==3.0.2
starlette==0.48.0
storage3==2.22.2
StrEnum==0.4.15
supabase==2.22.2
supabase-auth==2.22.2
supabase-functions==2.22.2
tldextract==5.3.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
websockets==15.0.1
yarl==1.22.0
